refactor(gui): apply MVC to reduce main_window and exercise_tab LOC (fixes #210)
